+++ AetherBags/changelog.md
@@ -4,6 +4,7 @@
 - Added custom per category sorting options.
 - Added an option to hide the fallback misc category label.
 - Added sorting options for the user, plugin, game and misc category display order.
+- Added options to control the window sizes for the various bag windows.
 
 # 1.0.0.2
 - Fix ghost hovers for expanded retainer grid too.